+++ sdk/servicebus/azure-servicebus/azure/servicebus/_common/message.py
@@ -17,7 +17,7 @@ from uamqp.constants import MessageState
 
 from .constants import (
     _BATCH_MESSAGE_OVERHEAD_COST,
-    ReceiveMode,
+    ServiceBusReceiveMode,
     _X_OPT_ENQUEUED_TIME,
     _X_OPT_SEQUENCE_NUMBER,
     _X_OPT_ENQUEUE_SEQUENCE_NUMBER,
@@ -545,10 +545,10 @@ class ServiceBusReceivedMessage(ServiceBusMessage):
             :caption: Checking the properties on a received message.
 
     """
-    def __init__(self, message, receive_mode=ReceiveMode.PeekLock, **kwargs):
-        # type: (uamqp.message.Message, ReceiveMode, Any) -> None
+    def __init__(self, message, receive_mode=ServiceBusReceiveMode.PeekLock, **kwargs):
+        # type: (uamqp.message.Message, ServiceBusReceiveMode, Any) -> None
         super(ServiceBusReceivedMessage, self).__init__(None, message=message)  # type: ignore
-        self._settled = (receive_mode == ReceiveMode.ReceiveAndDelete)
+        self._settled = (receive_mode == ServiceBusReceiveMode.ReceiveAndDelete)
         self._received_timestamp_utc = utc_now()
         self._is_deferred_message = kwargs.get("is_deferred_message", False)
         self._is_peeked_message = kwargs.get("is_peeked_message", False)