+++ sdk/servicebus/azure-servicebus/azure/servicebus/_servicebus_client.py
@@ -221,7 +221,7 @@ class ServiceBusClient(object):
          lock period before they will be removed from the queue. Messages received with ReceiveAndDelete
          will be immediately removed from the queue, and cannot be subsequently rejected or re-received if
          the client fails to process the message. The default receive_mode is PeekLock.
-        :paramtype receive_mode: ~azure.servicebus.ReceiveMode
+        :paramtype receive_mode: ~azure.servicebus.ServiceBusReceiveMode
         :keyword Optional[float] max_wait_time: The timeout in seconds between received messages after which the
          receiver will automatically stop receiving. The default value is None, meaning no timeout.
         :keyword Optional[~azure.servicebus.AutoLockRenewer] auto_lock_renewer: An ~azure.servicebus.AutoLockRenewer
@@ -330,7 +330,7 @@ class ServiceBusClient(object):
          lock period before they will be removed from the subscription. Messages received with ReceiveAndDelete
          will be immediately removed from the subscription, and cannot be subsequently rejected or re-received if
          the client fails to process the message. The default receive_mode is PeekLock.
-        :paramtype receive_mode: ~azure.servicebus.ReceiveMode
+        :paramtype receive_mode: ~azure.servicebus.ServiceBusReceiveMode
         :keyword Optional[float] max_wait_time: The timeout in seconds between received messages after which the
          receiver will automatically stop receiving. The default value is None, meaning no timeout.
         :keyword Optional[~azure.servicebus.AutoLockRenewer] auto_lock_renewer: An ~azure.servicebus.AutoLockRenewer