+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_client_async.py
@@ -213,7 +213,7 @@ class ServiceBusClient(object):
          lock period before they will be removed from the queue. Messages received with ReceiveAndDelete
          will be immediately removed from the queue, and cannot be subsequently rejected or re-received if
          the client fails to process the message. The default mode is PeekLock.
-        :paramtype receive_mode: ~azure.servicebus.ReceiveMode
+        :paramtype receive_mode: ~azure.servicebus.ServiceBusReceiveMode
         :keyword Optional[float] max_wait_time: The timeout in seconds between received messages after which the
          receiver will automatically stop receiving. The default value is None, meaning no timeout.
         :keyword Optional[~azure.servicebus.aio.AutoLockRenewer] auto_lock_renewer: An
@@ -319,7 +319,7 @@ class ServiceBusClient(object):
          lock period before they will be removed from the subscription. Messages received with ReceiveAndDelete
          will be immediately removed from the subscription, and cannot be subsequently rejected or re-received if
          the client fails to process the message. The default mode is PeekLock.
-        :paramtype receive_mode: ~azure.servicebus.ReceiveMode
+        :paramtype receive_mode: ~azure.servicebus.ServiceBusReceiveMode
         :keyword Optional[float] max_wait_time: The timeout in seconds between received messages after which the
          receiver will automatically stop receiving. The default value is None, meaning no timeout.
         :keyword Optional[~azure.servicebus.aio.AutoLockRenewer] auto_lock_renewer: An