+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_receiver_async.py
@@ -23,7 +23,7 @@ from .._common.constants import (
     REQUEST_RESPONSE_PEEK_OPERATION,
     REQUEST_RESPONSE_RECEIVE_BY_SEQUENCE_NUMBER,
     REQUEST_RESPONSE_RENEWLOCK_OPERATION,
-    ReceiveMode,
+    ServiceBusReceiveMode,
     MGMT_REQUEST_DISPOSITION_STATUS,
     MGMT_REQUEST_LOCK_TOKENS,
     MGMT_REQUEST_SEQUENCE_NUMBERS,
@@ -84,7 +84,7 @@ class ServiceBusReceiver(collections.abc.AsyncIterator, BaseHandler, ReceiverMix
      will be immediately removed from the queue, and cannot be subsequently abandoned or re-received
      if the client fails to process the message.
      The default mode is PeekLock.
-    :paramtype receive_mode: ~azure.servicebus.ReceiveMode
+    :paramtype receive_mode: ~azure.servicebus.ServiceBusReceiveMode
     :keyword Optional[float] max_wait_time: The timeout in seconds between received messages after which the receiver
      will automatically stop receiving. The default value is None, meaning no timeout.
     :keyword bool logging_enable: Whether to output network trace logs to the logger. Default is `False`.
@@ -205,7 +205,7 @@ class ServiceBusReceiver(collections.abc.AsyncIterator, BaseHandler, ReceiverMix
          will be immediately removed from the queue, and cannot be subsequently abandoned or re-received
          if the client fails to process the message.
          The default mode is PeekLock.
-        :paramtype receive_mode: ~azure.servicebus.ReceiveMode
+        :paramtype receive_mode: ~azure.servicebus.ServiceBusReceiveMode
         :keyword Optional[float] max_wait_time: The timeout in seconds between received messages after which the
          receiver will automatically stop receiving. The default value is None, meaning no timeout.
         :keyword bool logging_enable: Whether to output network trace logs to the logger. Default is `False`.
@@ -261,7 +261,9 @@ class ServiceBusReceiver(collections.abc.AsyncIterator, BaseHandler, ReceiverMix
             auto_complete=False,
             encoding=self._config.encoding,
             receive_settle_mode=self._receive_mode.value,
-            send_settle_mode=SenderSettleMode.Settled if self._receive_mode == ReceiveMode.ReceiveAndDelete else None,
+            send_settle_mode=SenderSettleMode.Settled \
+                if self._receive_mode == ServiceBusReceiveMode.ReceiveAndDelete \
+                else None,
             timeout=self._max_wait_time * 1000 if self._max_wait_time else 0,
             prefetch=self._prefetch_count,
             keep_alive_interval=self._config.keep_alive,