+++ sdk/servicebus/azure-servicebus/tests/async_tests/test_subscriptions_async.py
@@ -11,7 +11,7 @@ import pytest
 import time
 from datetime import datetime, timedelta
 
-from azure.servicebus import ServiceBusMessage, ReceiveMode
+from azure.servicebus import ServiceBusMessage, ServiceBusReceiveMode
 from azure.servicebus.aio import ServiceBusClient
 from azure.servicebus.aio._base_handler_async import ServiceBusSharedKeyCredential
 from azure.servicebus.exceptions import ServiceBusError
@@ -121,7 +121,7 @@ class ServiceBusSubscriptionAsyncTests(AzureMgmtTestCase):
                 topic_name=servicebus_topic.name,
                 subscription_name=servicebus_subscription.name,
                 max_wait_time=5,
-                receive_mode=ReceiveMode.PeekLock,
+                receive_mode=ServiceBusReceiveMode.PeekLock,
                 prefetch_count=10
             ) as receiver:
 
@@ -145,7 +145,7 @@ class ServiceBusSubscriptionAsyncTests(AzureMgmtTestCase):
                 topic_name=servicebus_topic.name,
                 subscription_name=servicebus_subscription.name,
                 max_wait_time=5,
-                receive_mode=ReceiveMode.PeekLock
+                receive_mode=ServiceBusReceiveMode.PeekLock
             ) as receiver:
                 count = 0
                 async for message in receiver:
@@ -159,7 +159,7 @@ class ServiceBusSubscriptionAsyncTests(AzureMgmtTestCase):
                 subscription_name=servicebus_subscription.name,
                 sub_queue = SubQueue.DeadLetter,
                 max_wait_time=5,
-                receive_mode=ReceiveMode.PeekLock
+                receive_mode=ServiceBusReceiveMode.PeekLock
             ) as dl_receiver:
                 count = 0
                 async for message in dl_receiver: