+++ sdk/servicebus/azure-servicebus/tests/test_queues.py
@@ -24,7 +24,7 @@ from azure.servicebus import (
     ServiceBusMessage,
     ServiceBusMessageBatch,
     ServiceBusReceivedMessage,
-    ReceiveMode,
+    ServiceBusReceiveMode,
     SubQueue
 )
 from azure.servicebus._common.constants import (
@@ -75,7 +75,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                 sender.send_messages(ServiceBusMessage("ServiceBusMessage {}".format(i)))
 
         with sb_client.get_queue_receiver(servicebus_queue.name,
-                                          receive_mode=ReceiveMode.ReceiveAndDelete,
+                                          receive_mode=ServiceBusReceiveMode.ReceiveAndDelete,
                                           max_wait_time=10) as receiver:
             batch = receiver.receive_messages()
             count = len(batch)
@@ -271,7 +271,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             messages = []
             with sb_client.get_queue_receiver(servicebus_queue.name, 
-                                              receive_mode=ReceiveMode.ReceiveAndDelete, 
+                                              receive_mode=ServiceBusReceiveMode.ReceiveAndDelete, 
                                               max_wait_time=8) as receiver:
                 for message in receiver:
                     assert not message.application_properties
@@ -298,7 +298,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             messages = []
             with sb_client.get_queue_receiver(servicebus_queue.name, 
-                                              receive_mode=ReceiveMode.ReceiveAndDelete, 
+                                              receive_mode=ServiceBusReceiveMode.ReceiveAndDelete, 
                                               max_wait_time=5) as receiver:
                 for message in receiver:
                     messages.append(message)
@@ -353,7 +353,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               max_wait_time=5, 
-                                              receive_mode=ReceiveMode.PeekLock) as receiver:
+                                              receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
             
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(10):
@@ -385,7 +385,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
         with ServiceBusClient.from_connection_string(
             servicebus_namespace_connection_string, logging_enable=False) as sb_client:
     
-            with sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=5, receive_mode=ReceiveMode.PeekLock) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=5, receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
             
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(10):
@@ -404,7 +404,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             assert count == 10
     
-            with sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=20, receive_mode=ReceiveMode.PeekLock) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=20, receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
                 count = 0
                 for message in receiver:
                     print_message(_logger, message)
@@ -427,7 +427,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
             with sb_client.get_queue_receiver(
                 servicebus_queue.name, 
                 max_wait_time=5, 
-                receive_mode=ReceiveMode.PeekLock) as receiver:
+                receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
             
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(10):
@@ -442,7 +442,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                     receiver.defer_message(message)
     
             assert count == 10
-            with sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=5, receive_mode=ReceiveMode.PeekLock) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, max_wait_time=5, receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
                 count = 0
                 for message in receiver:
                     print_message(_logger, message)
@@ -464,7 +464,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
             deferred_messages = []
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                                  max_wait_time=5, 
-                                                 receive_mode=ReceiveMode.PeekLock) as receiver:
+                                                 receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
             
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(10):
@@ -506,7 +506,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                         max_wait_time=5, 
-                                        receive_mode=ReceiveMode.PeekLock) as receiver:
+                                        receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
                 count = 0
                 for message in receiver:
                     deferred_messages.append(message.sequence_number)
@@ -518,7 +518,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                            max_wait_time=5, 
-                                           receive_mode=ReceiveMode.PeekLock) as receiver:
+                                           receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
                 deferred = receiver.receive_deferred_messages(deferred_messages)
                 assert len(deferred) == 10
                 for message in deferred:
@@ -545,7 +545,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               max_wait_time=5, 
-                                              receive_mode=ReceiveMode.PeekLock) as receiver:
+                                              receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
                 count = 0
                 for message in receiver:
                     deferred_messages.append(message.sequence_number)
@@ -601,7 +601,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             assert count == 10
             with sb_client.get_queue_receiver(servicebus_queue.name, 
-                                              receive_mode=ReceiveMode.ReceiveAndDelete, 
+                                              receive_mode=ServiceBusReceiveMode.ReceiveAndDelete, 
                                               max_wait_time=5) as receiver:
                 deferred = receiver.receive_deferred_messages(deferred_messages)
                 assert len(deferred) == 10
@@ -624,7 +624,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
             deferred_messages = []
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                                  max_wait_time=5, 
-                                                 receive_mode=ReceiveMode.PeekLock) as receiver:
+                                                 receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
             
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(3):
@@ -658,7 +658,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                            max_wait_time=5, 
-                                           receive_mode=ReceiveMode.PeekLock, 
+                                           receive_mode=ServiceBusReceiveMode.PeekLock, 
                                            prefetch_count=10) as receiver:
             
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
@@ -680,7 +680,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             with sb_client.get_queue_receiver(servicebus_queue.name,
                                               max_wait_time=5, 
-                                              receive_mode=ReceiveMode.PeekLock) as receiver:
+                                              receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
                 count = 0
                 for message in receiver:
                     print_message(_logger, message)
@@ -692,7 +692,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                     servicebus_queue.name,
                     sub_queue = SubQueue.DeadLetter,
                     max_wait_time=5,
-                    receive_mode=ReceiveMode.PeekLock) as dl_receiver:
+                    receive_mode=ServiceBusReceiveMode.PeekLock) as dl_receiver:
                 count = 0
                 for message in dl_receiver:
                     dl_receiver.complete_message(message)
@@ -715,7 +715,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             with sb_client.get_queue_receiver(servicebus_queue.name,
                                            max_wait_time=5,
-                                           receive_mode=ReceiveMode.PeekLock, 
+                                           receive_mode=ServiceBusReceiveMode.PeekLock, 
                                            prefetch_count=10) as receiver:
             
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
@@ -741,7 +741,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                     servicebus_queue.name,
                     sub_queue = SubQueue.DeadLetter,
                     max_wait_time=5,
-                    receive_mode=ReceiveMode.PeekLock) as dl_receiver:
+                    receive_mode=ServiceBusReceiveMode.PeekLock) as dl_receiver:
                 count = 0
                 for message in dl_receiver:
                     print_message(_logger, message)
@@ -807,7 +807,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
 
             receiver = sb_client.get_queue_receiver(servicebus_queue.name,
                                            max_wait_time=5,
-                                           receive_mode=ReceiveMode.PeekLock)
+                                           receive_mode=ServiceBusReceiveMode.PeekLock)
             sender = sb_client.get_queue_sender(servicebus_queue.name)
             with receiver, sender:
                 for i in range(5):
@@ -873,7 +873,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
     
             with sb_client.get_queue_receiver(servicebus_queue.name,
                                                  max_wait_time=5, 
-                                                 receive_mode=ReceiveMode.PeekLock, 
+                                                 receive_mode=ServiceBusReceiveMode.PeekLock, 
                                                  prefetch_count=10) as receiver:
                 messages = receiver.peek_messages(10)
                 assert len(messages) == 0
@@ -913,7 +913,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
             locks = 3
             with sb_client.get_queue_receiver(servicebus_queue.name,
                                               max_wait_time=5, 
-                                              receive_mode=ReceiveMode.PeekLock, 
+                                              receive_mode=ServiceBusReceiveMode.PeekLock, 
                                               prefetch_count=10) as receiver:
                 with sb_client.get_queue_sender(servicebus_queue.name) as sender:
                     for i in range(locks):
@@ -960,7 +960,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
             messages = []
             with sb_client.get_queue_receiver(servicebus_queue.name,
                                                  max_wait_time=5, 
-                                                 receive_mode=ReceiveMode.PeekLock, 
+                                                 receive_mode=ServiceBusReceiveMode.PeekLock, 
                                                  prefetch_count=10) as receiver:
                 for message in receiver:
                     if not messages:
@@ -1014,7 +1014,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
             messages = []
             with sb_client.get_queue_receiver(servicebus_queue.name,
                                                  max_wait_time=5, 
-                                                 receive_mode=ReceiveMode.PeekLock, 
+                                                 receive_mode=ServiceBusReceiveMode.PeekLock, 
                                                  prefetch_count=10,
                                                  auto_lock_renewer=renewer) as receiver:
                 for message in receiver:
@@ -1074,7 +1074,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                     servicebus_queue.name,
                     sub_queue = SubQueue.DeadLetter,
                     max_wait_time=5,
-                    receive_mode=ReceiveMode.PeekLock) as dl_receiver:
+                    receive_mode=ServiceBusReceiveMode.PeekLock) as dl_receiver:
                 count = 0
                 for message in dl_receiver:
                     print_message(_logger, message)
@@ -1213,7 +1213,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                 sender.send_messages(message)
     
             with sb_client.get_queue_receiver(servicebus_queue.name,
-                                                 receive_mode=ReceiveMode.ReceiveAndDelete) as receiver:
+                                                 receive_mode=ServiceBusReceiveMode.ReceiveAndDelete) as receiver:
                 messages = receiver.receive_messages(max_wait_time=10)
                 assert len(messages) == 1
                 message = messages[0]
@@ -1450,7 +1450,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
                 message = ServiceBusMessage("Test")
                 sender.send_messages(message)
 
-            with sb_client.get_queue_receiver(servicebus_queue.name, receive_mode=ReceiveMode.ReceiveAndDelete) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, receive_mode=ServiceBusReceiveMode.ReceiveAndDelete) as receiver:
                 assert receiver._config.transport_type == TransportType.AmqpOverWebsocket
                 messages = receiver.receive_messages(max_wait_time=5)
                 assert len(messages) == 1
@@ -1972,7 +1972,7 @@ class ServiceBusQueueTests(AzureMgmtTestCase):
             servicebus_namespace_connection_string, logging_enable=False) as sb_client:
             with pytest.raises(ValueError):
                 with sb_client.get_queue_receiver(servicebus_queue.name, 
-                                                  receive_mode=ReceiveMode.ReceiveAndDelete,
+                                                  receive_mode=ServiceBusReceiveMode.ReceiveAndDelete,
                                                   auto_lock_renewer=AutoLockRenewer()) as receiver:
                 
                     raise Exception("Should not get here, should fail fast because ReceiveAndDelete messages cannot be autorenewed.")