+++ sdk/servicebus/azure-servicebus/tests/test_sessions.py
@@ -18,7 +18,7 @@ from azure.servicebus import (
     AutoLockRenewer,
     ServiceBusMessage,
     ServiceBusReceivedMessage,
-    ReceiveMode,
+    ServiceBusReceiveMode,
     NEXT_AVAILABLE_SESSION,
     SubQueue
 )
@@ -124,7 +124,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             messages = []
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
-                                              receive_mode=ReceiveMode.ReceiveAndDelete, 
+                                              receive_mode=ServiceBusReceiveMode.ReceiveAndDelete, 
                                               max_wait_time=5) as receiver:
                 for message in receiver:
                     messages.append(message)
@@ -138,7 +138,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             time.sleep(30)
 
             messages = []
-            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ReceiveMode.ReceiveAndDelete, max_wait_time=5) as session:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, receive_mode=ServiceBusReceiveMode.ReceiveAndDelete, max_wait_time=5) as session:
                 for message in session:
                     messages.append(message)
                 assert len(messages) == 0
@@ -249,7 +249,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             messages = []
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
-                                              receive_mode=ReceiveMode.ReceiveAndDelete, 
+                                              receive_mode=ServiceBusReceiveMode.ReceiveAndDelete, 
                                               max_wait_time=5) as session:
                 for message in session:
                     messages.append(message)
@@ -378,7 +378,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             with sb_client.get_queue_receiver(servicebus_queue.name, 
                                               session_id=session_id, 
                                               max_wait_time=5,
-                                              receive_mode=ReceiveMode.ReceiveAndDelete) as receiver:
+                                              receive_mode=ServiceBusReceiveMode.ReceiveAndDelete) as receiver:
                 deferred = receiver.receive_deferred_messages(deferred_messages)
                 assert len(deferred) == 10
                 for message in deferred:
@@ -594,7 +594,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
 
             renewer = AutoLockRenewer()
             messages = []
-            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=10) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ServiceBusReceiveMode.PeekLock, prefetch_count=10) as receiver:
                 renewer.register(receiver,
                                  receiver.session,
                                  max_lock_renewal_duration=10,
@@ -635,7 +635,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             renewer._renew_period = 1
             session = None
 
-            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ReceiveMode.PeekLock, prefetch_count=10) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id, max_wait_time=5, receive_mode=ServiceBusReceiveMode.PeekLock, prefetch_count=10) as receiver:
                 session = receiver.session
                 renewer.register(receiver,
                                  session,
@@ -672,7 +672,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             with sb_client.get_queue_receiver(servicebus_queue.name,
                                               session_id=session_id,
                                               max_wait_time=5,
-                                              receive_mode=ReceiveMode.PeekLock,
+                                              receive_mode=ServiceBusReceiveMode.PeekLock,
                                               prefetch_count=10,
                                               auto_lock_renewer=renewer) as receiver:
                 print("Registered lock renew thread", receiver.session._locked_until_utc, utc_now())
@@ -714,7 +714,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
             with sb_client.get_queue_receiver(servicebus_queue.name,
                                               session_id=session_id,
                                               max_wait_time=5,
-                                              receive_mode=ReceiveMode.PeekLock,
+                                              receive_mode=ServiceBusReceiveMode.PeekLock,
                                               prefetch_count=10,
                                               auto_lock_renewer=renewer) as receiver:
                 session = receiver.session
@@ -947,7 +947,7 @@ class ServiceBusSessionTests(AzureMgmtTestCase):
                 with sb_client.get_queue_receiver(servicebus_queue.name, session_id=session_id) as receiver:
                     receiver.set_state("SESSION {}".format(session_id))
 
-            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION, max_wait_time=5, receive_mode=ReceiveMode.PeekLock) as receiver:
+            with sb_client.get_queue_receiver(servicebus_queue.name, session_id=NEXT_AVAILABLE_SESSION, max_wait_time=5, receive_mode=ServiceBusReceiveMode.PeekLock) as receiver:
                 current_sessions = receiver.list_sessions(updated_since=start_time)
                 assert len(current_sessions) == 5
                 assert current_sessions == sessions